Run pre-commit in the lint workflow (#280)
Introduce 3 versions of pre-commit, for three versions of python: 3.10, 3.11, 3.12. This is due to the observed behavior of `flake8` giving different results with different versions of Python. Only pre-commit with Python 3.10 is required, and the others pass on error but are included for information and insight.

All the other changes are for fixing the problems flagged by pre-commit.