Decrease address randomization to pass sanitizer tests (#629)
Libsanitizer has a bug where tests are not reliable with high address randomization. This PR decreases address randomization for sanitizer tests. It goes along with a change in the nv-legate/legate-gh-ci that runs containers in the privileged mode. This fix can be removed when new version of sanitizer comes out and fixes this bug.