Install the Legion libraries in lib with a suffix (#1741)
* Install the Legion libraries in lib with a suffix

This enables us to append a suffix to the installed library names, they
are stored in the installed target description and mitigate name clashes
while also allowing the library to be used in the `INTERFACE`.

* Add legion and realm suppression with -legate suffix

Adding rather than replacing to avoid issues if we build with or without
the suffix on the Legion library names on disk.