Build python pip binary wheels (#1822)
Add a Python binary pip wheel build to the project. This builds against system compilers on Rocky Linux 8 to achieve a highly compatible ABI that will run on a large cross section of Linux hosts. In addition to CPM/rapids-cmake vendored libraries the wheels currently build and graft in a recent HDF5 library version.

The binary layout of the wheels is quite different to conda. Where conda typically locates the Python package in `lib/python3.12/site-packages/legate` with all C++ artifacts in `lib` the Python wheels are for the most part self-contained with the Python package in the same place but the C++ artifacts are in `lib/python3.12/site-packages/legate/lib64`.

Longer term the plan is to split the C++ artifacts into `liblegate` for all of Python 3.x as they are the largest and require the most build time. We will then put all Python (including cython generated artifacts) in `legate` that depends upon `liblegate`. The `cupynumeric` wheel downloads and compiles against this wheel, linking to it at runtime.

The wheels are currently built for PRs, that needs some attention, but at this point I would like to get review on the build and Python changes. In parallel I will look at trimmed down builds for PRs (maybe two architectures, just one Python version).