Build the MPI wrappers as part of the pip wheel (#1977)
Prebuild the MPI wrappers for the legate wheels against OpenMPI and MPICH, and add them to `lib64/legate/wrappers` with an additional RPATH to help `dlopen` find them. This vendors the `realm` MPI bootstrap code (and installs it for optional user compilation), and reuses the `legate` MPI wrapper code.

The wrappers produced have had some testing on other systems and worked in those cases.