Use the UCX wheel library loader and prefer system (#2040)
The UCX wheel library loader has logic to attempt to prefer the system UCX over the bundled wheels in pip. Add some logic to make the init aware of being in a wheel build and perform some custom logic for loading UCX before any other libraries are loaded.

Thanks to Vyas Ramasubramani and Jacob Faibussowitsch for advice on putting this together.