Add a source build to CI using a vanilla container (#2468)
Add a source build to the CI matrix and parameterize the GCC major version. Compress the build tree, filtering out files we don't need to save space, and run the tests in a GPU container. Also switch from the `devel` container for build to the `base` for testing. This offers a much smaller test container speeding up the workflow.

Filter out some of the tests that cannot be simply run by `ctest`, these should be addressed in a follow up. This new workflow builds and tests a typical build from source on Rocky Linux 8 outside of the PyPi or conda ecosystem more typical of HPC deployments.