Build binary wheels for mac (#2919)
Note that the mac workflow diverges and cannot/should not share the Linux workflow. It uses a self-hosted runner, no container, and is CPU only so build and testing can happen in the same job. This builds and uploads the binary mac wheels and leaves testing for a follow up PR.