Remove the version from Legion and Realm finds (#3141)
This makes it challenging to build against source trees locally, and in normal builds we already specify the commit SHA to build against. Set to 0.0.0 when a custom source directory was supplied, otherwise keep the version so that it is a little clearer in configure what the version being built should (loosely) correspond to.