Add very simple testing for mac (#3303)
This adds back the full matrix and then applies a filter to it for the latest Python version. It adds a simple smoke test that would have caught the last regression, and avoids adding in the full test suite as it required HDF5 support and MPI. Increasing the timeout for the job as there are sporadic issues with the large git repositories cloned during configure.